Take screenshot on failure and upon request
